requirements.txt
flask==3.1.3
flask-cors==6.0.2
gunicorn==25.3.0
yfinance==1.3.0
feedparser==6.0.12
requests==2.33.1
groq==1.2.0
anthropic==0.97.0
sentence-transformers==5.4.1
numpy==2.4.4
psycopg2-binary==2.9.12
